추가 수정 5. view count 타입 수정 : Long -> Integer 6. 브랜드 조회수 증가 기능 반환값 변경 : null -> Integer

+++ src/main/java/com/ward/ward_server/api/item/dto/TopBrandResponse.java
@@ -5,8 +5,8 @@ import java.util.List;
 public record TopBrandResponse(
         String brandLogoImage,
         String brandName,
-        Long brandViewCount,
-        Long brandWishCount,
+        Integer brandViewCount,
+        Integer brandWishCount,
         List<TopBrandItemResponse> top3itemList
 ) {
 }
